refatoração e adição das us 9 e 10 junto com as 6 e 7

+++ src/main/java/com/psoft/wallet/service/NotificationService.java
@@ -3,13 +3,14 @@ package com.psoft.wallet.service;
 import com.psoft.wallet.model.Ativo;
 import com.psoft.wallet.model.Cliente;
 import org.springframework.stereotype.Service;
+import java.math.BigDecimal;
 
 @Service
 public class NotificationService {
 
-    public void notificarVariacaoPreco(Cliente cliente, Ativo ativo, float precoAntigo, float precoNovo) {
+    public void notificarVariacaoPreco(Cliente cliente, Ativo ativo, BigDecimal precoAntigo, BigDecimal precoNovo) {
         System.out.printf(
-                "NOTIFICAÇÃO para %s: O ativo '%s' teve uma variação de preço significativa. Valor anterior: %.1f, Valor atual: %.1f.%n",
+                "NOTIFICAÇÃO para %s: O ativo '%s' teve uma variação de preço significativa. Valor anterior: %.2f, Valor atual: %.2f.%n",
                 cliente.getNomeCompleto(), ativo.getNome(), precoAntigo, precoNovo
         );
     }
